Added dataset used for crime_analysis R report
Crime.csv file has two columns, one of which is the percentage of individuals in a county with at least a high-school diploma (columndip), and the other is the crime rate per 100,000 residents for the counties (columnrate).  Y is considered to by crime rate and X the percentage of individuals with at least a high-school diploma.